renamed dir to CLI and GUI
