+++ src/CLI/inventory.py
